// src/model/area/areaKey.ts
export enum AreaKey {
    TheStonyShore = 'TheStonyShore',
    Winterfell = 'Winterfell',
    WhiteHarbor = 'WhiteHarbor',
    CastleBlack = 'CastleBlack',
    Karhold = 'Karhold',
    WidowsWatch = 'WidowsWatch',
    FlintsFinger = 'FlintsFinger',
    GreyWaterWatch = 'GreyWaterWatch',
    MoatCailin = 'MoatCailin',
    Seagard = 'Seagard',
    TheTwins = 'TheTwins',
    TheFingers = 'TheFingers',
    Lannisport = 'Lannisport',
    TheEyrie = 'TheEyrie',
    TheNarrowSea = 'TheNarrowSea',
    BayOfIce = 'BayOfIce',
    TheShiveringSea = 'TheShiveringSea',
    SunsetSea = 'SunsetSea',
    IronmansBay = 'IronmansBay',
    Pyke = 'Pyke',
    TheGoldenSound = 'TheGoldenSound',
    Riverrun = 'Riverrun',
    TheMountainsOfTheMoon = 'TheMountainsOfTheMoon',
    Harrenhal = 'Harrenhal',
    StoneySept = 'StoneySept',
    CrackClawPoint = 'CrackClawPoint',
    Dragonstone = 'Dragonstone',
    BlackwaterBay = 'BlackwaterBay',
    ShipbreakerBay = 'ShipbreakerBay',
    Kingswood = 'Kingswood',
    KingsLanding = 'KingsLanding',
    Blackwater = 'Blackwater',
    SearoadMarches = 'SearoadMarches',
    Highgarden = 'Highgarden',
    StormsEnd = 'StormsEnd',
    TheBoneway = 'TheBoneway',
    DornishMarches = 'DornishMarches',
    Oldtown = 'Oldtown',
    ThreeTowers = 'ThreeTowers',
    PrincesPass = 'PrincesPass',
    Yronwood = 'Yronwood',
    Sunspear = 'Sunspear',
    SaltShore = 'SaltShore',
    Starfall = 'Starfall',
    SeaOfDorne = 'SeaOfDorne',
    TheArbor = 'TheArbor',
    RedwyneStraights = 'RedwyneStraights',
    WestSummerSea = 'WestSummerSea',
    EastSummerSea = 'EastSummerSea',
    TheReach = 'TheReach'
}